Restructure autogenerated application to conform to app_ syntax
